Initial commit of code
This includes a (mostly) working version of the ISP trajectory/
viewing geometry code I wrote for the workshop last year. Fair
warning, I've tried to make it as general as possible, but there
are likely to be hidden hardcodings. I've also had to remove the
skyfield data files, and will need to add a how-to to get the ones
I've been using to the repo at some point.